refactor: rename src/ to components/
Flatten the pre-existing src/components/ui/ into components/ui/ so
the rename doesn't produce a components/components/ nesting, and
update every import path (entrypoints, .ladle config, package.json
scripts) accordingly.
