+++ .ladle/components.tsx
@@ -1,4 +1,4 @@
-import "../src/globals.css";
+import "../components/globals.css";
 import React from "react";
 
 export const Provider = ({ children }: { children: React.ReactNode }) => (
