+++ .ladle/config.mjs
@@ -1,5 +1,5 @@
 /** @type {import('@ladle/react').UserConfig} */
 export default {
-  stories: "src/**/*.stories.{js,jsx,ts,tsx}",
+  stories: "components/**/*.stories.{js,jsx,ts,tsx}",
   viteConfig: "./vite.config.ts",
 };
