+++ entrypoints/options/main.tsx
@@ -1,7 +1,7 @@
-import "../../src/globals.css";
+import "../../components/globals.css";
 import React from "react";
 import ReactDOM from "react-dom/client";
-import { Options } from "../../src/options/Options";
+import { Options } from "../../components/options/Options";
 
 ReactDOM.createRoot(document.getElementById("root")!).render(
   <React.StrictMode>
